TFP-5999: Sammenfatter ulike implementasjonene av steg for vurdere kompletthet og søkt for tidlig
