+++ behandlingsprosess/src/main/java/no/nav/foreldrepenger/behandling/steg/kompletthet/VurderKompletthetStegFelles.java
@@ -6,13 +6,15 @@ import java.time.LocalDateTime;
 
 import no.nav.foreldrepenger.behandlingskontroll.BehandleStegResultat;
 import no.nav.foreldrepenger.behandlingslager.behandling.Behandling;
+import no.nav.foreldrepenger.behandlingslager.behandling.BehandlingÅrsak;
+import no.nav.foreldrepenger.behandlingslager.behandling.BehandlingÅrsakType;
 import no.nav.foreldrepenger.behandlingslager.behandling.aksjonspunkt.Aksjonspunkt;
 import no.nav.foreldrepenger.behandlingslager.behandling.aksjonspunkt.AksjonspunktDefinisjon;
 import no.nav.foreldrepenger.kompletthet.KompletthetResultat;
 
 /**
  * Fellesklasse for gjenbrukte metode av subklasser for
- * {@link VurderKompletthetSteg}.
+ * {@link VurderKompletthetSteg} og {@link VurderSøktForTidligSteg}.
  * <p>
  * Favor composition over inheritance
  */
@@ -37,5 +39,17 @@ public class VurderKompletthetStegFelles {
         return behandling.getAksjonspunktMedDefinisjonOptional(apDef)
                 .map(Aksjonspunkt::erUtført)
                 .orElse(Boolean.FALSE);
+    }
+
+    public static boolean skalPassereKompletthet(Behandling behandling) {
+        return behandling.getBehandlingÅrsaker().stream()
+            .map(BehandlingÅrsak::getBehandlingÅrsakType)
+            .anyMatch(BehandlingÅrsakType.årsakerRelatertTilDød()::contains);
+    }
+
+    public static boolean kanPassereKompletthet(Behandling behandling) {
+        return behandling.getBehandlingÅrsaker().stream()
+            .map(BehandlingÅrsak::getBehandlingÅrsakType)
+            .anyMatch(BehandlingÅrsakType.RE_HENDELSE_FØDSEL::equals);
     }
 }
