+++ behandlingsprosess/src/main/java/no/nav/foreldrepenger/behandling/steg/kompletthet/VurderSøktForTidligSteg.java
@@ -1,5 +1,7 @@
-package no.nav.foreldrepenger.behandling.steg.kompletthet.fp;
+package no.nav.foreldrepenger.behandling.steg.kompletthet;
 
+import static no.nav.foreldrepenger.behandling.steg.kompletthet.VurderKompletthetStegFelles.kanPassereKompletthet;
+import static no.nav.foreldrepenger.behandling.steg.kompletthet.VurderKompletthetStegFelles.skalPassereKompletthet;
 import static no.nav.foreldrepenger.behandlingslager.behandling.aksjonspunkt.AksjonspunktDefinisjon.VENT_PGA_FOR_TIDLIG_SØKNAD;
 
 import java.time.LocalDate;
@@ -10,16 +12,12 @@ import jakarta.inject.Inject;
 import no.nav.foreldrepenger.behandling.BehandlingReferanse;
 import no.nav.foreldrepenger.behandling.FamilieHendelseDato;
 import no.nav.foreldrepenger.behandling.Skjæringstidspunkt;
-import no.nav.foreldrepenger.behandling.steg.kompletthet.VurderKompletthetSteg;
-import no.nav.foreldrepenger.behandling.steg.kompletthet.VurderKompletthetStegFelles;
 import no.nav.foreldrepenger.behandlingskontroll.BehandleStegResultat;
+import no.nav.foreldrepenger.behandlingskontroll.BehandlingSteg;
 import no.nav.foreldrepenger.behandlingskontroll.BehandlingStegRef;
-import no.nav.foreldrepenger.behandlingskontroll.BehandlingTypeRef;
 import no.nav.foreldrepenger.behandlingskontroll.BehandlingskontrollKontekst;
-import no.nav.foreldrepenger.behandlingskontroll.FagsakYtelseTypeRef;
 import no.nav.foreldrepenger.behandlingslager.behandling.Behandling;
 import no.nav.foreldrepenger.behandlingslager.behandling.BehandlingStegType;
-import no.nav.foreldrepenger.behandlingslager.behandling.BehandlingType;
 import no.nav.foreldrepenger.behandlingslager.behandling.personopplysning.RelasjonsRolleType;
 import no.nav.foreldrepenger.behandlingslager.behandling.repository.BehandlingRepository;
 import no.nav.foreldrepenger.behandlingslager.behandling.repository.BehandlingRepositoryProvider;
@@ -33,23 +31,21 @@ import no.nav.foreldrepenger.skjæringstidspunkt.SkjæringstidspunktTjeneste;
 import no.nav.fpsak.tidsserie.LocalDateInterval;
 
 @BehandlingStegRef(BehandlingStegType.VURDER_KOMPLETT_TIDLIG)
-@BehandlingTypeRef(BehandlingType.FØRSTEGANGSSØKNAD)
-@FagsakYtelseTypeRef(FagsakYtelseType.FORELDREPENGER)
 @ApplicationScoped
-public class VurderSøktForTidligStegImpl implements VurderKompletthetSteg {
+public class VurderSøktForTidligSteg implements BehandlingSteg {
 
     private Kompletthetsjekker kompletthetsjekker;
     private BehandlingRepository behandlingRepository;
     private YtelsesFordelingRepository ytelsesFordelingRepository;
     private SkjæringstidspunktTjeneste skjæringstidspunktTjeneste;
 
-    VurderSøktForTidligStegImpl() {
+    VurderSøktForTidligSteg() {
     }
 
     @Inject
-    public VurderSøktForTidligStegImpl(Kompletthetsjekker kompletthetsjekker,
-                                       BehandlingRepositoryProvider provider,
-                                       SkjæringstidspunktTjeneste skjæringstidspunktTjeneste) {
+    public VurderSøktForTidligSteg(Kompletthetsjekker kompletthetsjekker,
+                                   BehandlingRepositoryProvider provider,
+                                   SkjæringstidspunktTjeneste skjæringstidspunktTjeneste) {
         this.kompletthetsjekker = kompletthetsjekker;
         this.skjæringstidspunktTjeneste = skjæringstidspunktTjeneste;
         this.behandlingRepository = provider.getBehandlingRepository();
@@ -58,27 +54,23 @@ public class VurderSøktForTidligStegImpl implements VurderKompletthetSteg {
 
     @Override
     public BehandleStegResultat utførSteg(BehandlingskontrollKontekst kontekst) {
-        var behandlingId = kontekst.getBehandlingId();
-        var behandling = behandlingRepository.hentBehandling(behandlingId);
-        var skjæringstidspunkter = skjæringstidspunktTjeneste.getSkjæringstidspunkter(behandlingId);
-        var ref = BehandlingReferanse.fra(behandling);
-
-        if (skalPassereKompletthet(behandling)) {
+        var behandling = behandlingRepository.hentBehandling(kontekst.getBehandlingId());
+        if (skalPassereKompletthet(behandling) || behandling.erRevurdering() || kanPassereKompletthet(behandling)) {
             return BehandleStegResultat.utførtUtenAksjonspunkter();
         }
 
-        var søknadMottatt = kompletthetsjekker.vurderSøknadMottattForTidlig(ref, skjæringstidspunkter);
-        if (!søknadMottatt.erOppfylt() && skalVenteDersomSøktForTidlig(behandling, skjæringstidspunkter)) {
+        var skjæringstidspunkter = skjæringstidspunktTjeneste.getSkjæringstidspunkter(kontekst.getBehandlingId());
+        var søknadMottatt = kompletthetsjekker.vurderSøknadMottattForTidlig(BehandlingReferanse.fra(behandling), skjæringstidspunkter);
+        if (søknadMottatt.erOppfylt()) {
+            return BehandleStegResultat.utførtUtenAksjonspunkter();
+        }
+        if (FagsakYtelseType.FORELDREPENGER.equals(behandling.getFagsakYtelseType()) && skalVenteDersomSøktForTidlig(behandling, skjæringstidspunkter)) {
             return VurderKompletthetStegFelles.evaluerUoppfylt(søknadMottatt, VENT_PGA_FOR_TIDLIG_SØKNAD);
         }
-
         return BehandleStegResultat.utførtUtenAksjonspunkter();
     }
 
     private boolean skalVenteDersomSøktForTidlig(Behandling behandling, Skjæringstidspunkt stp) {
-        if (!kanPassereKompletthet(behandling)) {
-            return true;
-        }
         if (RelasjonsRolleType.erMor(behandling.getRelasjonsRolleType())) {
             return false;
         }
