--- behandlingsprosess/src/main/java/no/nav/foreldrepenger/behandling/steg/kompletthet/es/VurderKompletthetStegESImpl.java
@@ -1,61 +1,0 @@
-package no.nav.foreldrepenger.behandling.steg.kompletthet.es;
-
-import static no.nav.foreldrepenger.behandlingslager.behandling.aksjonspunkt.AksjonspunktDefinisjon.AUTO_VENTER_PÅ_KOMPLETT_SØKNAD;
-
-import jakarta.enterprise.context.ApplicationScoped;
-import jakarta.inject.Inject;
-
-import no.nav.foreldrepenger.behandling.BehandlingReferanse;
-import no.nav.foreldrepenger.behandling.steg.kompletthet.VurderKompletthetSteg;
-import no.nav.foreldrepenger.behandling.steg.kompletthet.VurderKompletthetStegFelles;
-import no.nav.foreldrepenger.behandlingskontroll.BehandleStegResultat;
-import no.nav.foreldrepenger.behandlingskontroll.BehandlingStegRef;
-import no.nav.foreldrepenger.behandlingskontroll.BehandlingTypeRef;
-import no.nav.foreldrepenger.behandlingskontroll.BehandlingskontrollKontekst;
-import no.nav.foreldrepenger.behandlingskontroll.FagsakYtelseTypeRef;
-import no.nav.foreldrepenger.behandlingslager.behandling.BehandlingStegType;
-import no.nav.foreldrepenger.behandlingslager.behandling.repository.BehandlingRepository;
-import no.nav.foreldrepenger.behandlingslager.behandling.repository.BehandlingRepositoryProvider;
-import no.nav.foreldrepenger.behandlingslager.fagsak.FagsakYtelseType;
-import no.nav.foreldrepenger.kompletthet.Kompletthetsjekker;
-import no.nav.foreldrepenger.skjæringstidspunkt.SkjæringstidspunktTjeneste;
-
-@BehandlingStegRef(BehandlingStegType.VURDER_KOMPLETT_BEH)
-@BehandlingTypeRef
-@FagsakYtelseTypeRef(FagsakYtelseType.ENGANGSTØNAD)
-@ApplicationScoped
-public class VurderKompletthetStegESImpl implements VurderKompletthetSteg {
-
-    private Kompletthetsjekker vurderKompletthetTjeneste;
-    private BehandlingRepository behandlingRepository;
-    private SkjæringstidspunktTjeneste skjæringstidspunktTjeneste;
-
-    VurderKompletthetStegESImpl() {
-    }
-
-    @Inject
-    public VurderKompletthetStegESImpl(Kompletthetsjekker vurderKompletthetTjeneste,
-                                       BehandlingRepositoryProvider provider,
-                                       SkjæringstidspunktTjeneste skjæringstidspunktTjeneste) {
-        this.vurderKompletthetTjeneste = vurderKompletthetTjeneste;
-        this.skjæringstidspunktTjeneste = skjæringstidspunktTjeneste;
-        this.behandlingRepository = provider.getBehandlingRepository();
-    }
-
-    @Override
-    public BehandleStegResultat utførSteg(BehandlingskontrollKontekst kontekst) {
-        var behandling = behandlingRepository.hentBehandling(kontekst.getBehandlingId());
-        var skjæringstidspunkter = skjæringstidspunktTjeneste.getSkjæringstidspunkter(kontekst.getBehandlingId());
-        var ref = BehandlingReferanse.fra(behandling);
-
-        if (skalPassereKompletthet(behandling)) {
-            return BehandleStegResultat.utførtUtenAksjonspunkter();
-        }
-
-        var kompletthetResultat = vurderKompletthetTjeneste.vurderForsendelseKomplett(ref, skjæringstidspunkter);
-        if (!kompletthetResultat.erOppfylt() && !VurderKompletthetStegFelles.autopunktAlleredeUtført(AUTO_VENTER_PÅ_KOMPLETT_SØKNAD, behandling)) {
-            return VurderKompletthetStegFelles.evaluerUoppfylt(kompletthetResultat, AUTO_VENTER_PÅ_KOMPLETT_SØKNAD);
-        }
-        return BehandleStegResultat.utførtUtenAksjonspunkter();
-    }
-}
