--- behandlingsprosess/src/main/java/no/nav/foreldrepenger/behandling/steg/kompletthet/es/VurderSøktForTidligStegESImpl.java
@@ -1,30 +1,0 @@
-package no.nav.foreldrepenger.behandling.steg.kompletthet.es;
-
-import jakarta.enterprise.context.ApplicationScoped;
-import jakarta.inject.Inject;
-
-import no.nav.foreldrepenger.behandling.steg.kompletthet.VurderKompletthetSteg;
-import no.nav.foreldrepenger.behandlingskontroll.BehandleStegResultat;
-import no.nav.foreldrepenger.behandlingskontroll.BehandlingStegRef;
-import no.nav.foreldrepenger.behandlingskontroll.BehandlingTypeRef;
-import no.nav.foreldrepenger.behandlingskontroll.BehandlingskontrollKontekst;
-import no.nav.foreldrepenger.behandlingskontroll.FagsakYtelseTypeRef;
-import no.nav.foreldrepenger.behandlingslager.behandling.BehandlingStegType;
-import no.nav.foreldrepenger.behandlingslager.fagsak.FagsakYtelseType;
-
-@BehandlingStegRef(BehandlingStegType.VURDER_KOMPLETT_TIDLIG)
-@BehandlingTypeRef
-@FagsakYtelseTypeRef(FagsakYtelseType.ENGANGSTØNAD)
-@ApplicationScoped
-public class VurderSøktForTidligStegESImpl implements VurderKompletthetSteg {
-
-    @Inject
-    public VurderSøktForTidligStegESImpl() {
-        // Intentionally left blank
-    }
-
-    @Override
-    public BehandleStegResultat utførSteg(BehandlingskontrollKontekst kontekst) {
-        return BehandleStegResultat.utførtUtenAksjonspunkter();
-    }
-}
