--- behandlingsprosess/src/main/java/no/nav/foreldrepenger/behandling/steg/kompletthet/fp/VurderKompletthetStegImpl.java
@@ -1,72 +1,0 @@
-package no.nav.foreldrepenger.behandling.steg.kompletthet.fp;
-
-import static no.nav.foreldrepenger.behandlingslager.behandling.aksjonspunkt.AksjonspunktDefinisjon.AUTO_VENTER_PÅ_KOMPLETT_SØKNAD;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-
-import jakarta.enterprise.context.ApplicationScoped;
-import jakarta.inject.Inject;
-
-import no.nav.foreldrepenger.behandling.BehandlingReferanse;
-import no.nav.foreldrepenger.behandling.steg.kompletthet.VurderKompletthetSteg;
-import no.nav.foreldrepenger.behandling.steg.kompletthet.VurderKompletthetStegFelles;
-import no.nav.foreldrepenger.behandlingskontroll.BehandleStegResultat;
-import no.nav.foreldrepenger.behandlingskontroll.BehandlingStegRef;
-import no.nav.foreldrepenger.behandlingskontroll.BehandlingTypeRef;
-import no.nav.foreldrepenger.behandlingskontroll.BehandlingskontrollKontekst;
-import no.nav.foreldrepenger.behandlingskontroll.FagsakYtelseTypeRef;
-import no.nav.foreldrepenger.behandlingslager.behandling.BehandlingStegType;
-import no.nav.foreldrepenger.behandlingslager.behandling.BehandlingType;
-import no.nav.foreldrepenger.behandlingslager.behandling.repository.BehandlingRepository;
-import no.nav.foreldrepenger.behandlingslager.behandling.repository.BehandlingRepositoryProvider;
-import no.nav.foreldrepenger.behandlingslager.fagsak.FagsakYtelseType;
-import no.nav.foreldrepenger.kompletthet.Kompletthetsjekker;
-import no.nav.foreldrepenger.skjæringstidspunkt.SkjæringstidspunktTjeneste;
-
-@BehandlingStegRef(BehandlingStegType.VURDER_KOMPLETT_BEH)
-@BehandlingTypeRef(BehandlingType.FØRSTEGANGSSØKNAD)
-@FagsakYtelseTypeRef(FagsakYtelseType.FORELDREPENGER)
-@ApplicationScoped
-public class VurderKompletthetStegImpl implements VurderKompletthetSteg {
-
-    private Kompletthetsjekker kompletthetsjekker;
-    private BehandlingRepository behandlingRepository;
-    private SkjæringstidspunktTjeneste skjæringstidspunktTjeneste;
-
-    VurderKompletthetStegImpl() {
-    }
-
-    @Inject
-    public VurderKompletthetStegImpl(Kompletthetsjekker kompletthetsjekker,
-                                     BehandlingRepositoryProvider provider,
-                                     SkjæringstidspunktTjeneste skjæringstidspunktTjeneste) {
-        this.kompletthetsjekker = kompletthetsjekker;
-        this.skjæringstidspunktTjeneste = skjæringstidspunktTjeneste;
-        this.behandlingRepository = provider.getBehandlingRepository();
-    }
-
-    @Override
-    public BehandleStegResultat utførSteg(BehandlingskontrollKontekst kontekst) {
-        var behandlingId = kontekst.getBehandlingId();
-        var behandling = behandlingRepository.hentBehandling(behandlingId);
-        var skjæringstidspunkter = skjæringstidspunktTjeneste.getSkjæringstidspunkter(behandlingId);
-        var ref = BehandlingReferanse.fra(behandling);
-
-        if (skalPassereKompletthet(behandling)) {
-            return BehandleStegResultat.utførtUtenAksjonspunkter();
-        }
-
-        var forsendelseMottatt = kompletthetsjekker.vurderForsendelseKomplett(ref, skjæringstidspunkter);
-        if (!forsendelseMottatt.erOppfylt() && !VurderKompletthetStegFelles.autopunktAlleredeUtført(AUTO_VENTER_PÅ_KOMPLETT_SØKNAD, behandling)) {
-            // kompletthetsresultat kan være langt fram i tid dersom tidlig fødsel
-            var brukfrist = kanPassereKompletthet(behandling) && !forsendelseMottatt.erFristUtløpt()
-                && forsendelseMottatt.ventefrist().isAfter(LocalDateTime.now().plusWeeks(1)) ?
-                LocalDate.now().plusWeeks(1).atStartOfDay() : forsendelseMottatt.ventefrist();
-            return VurderKompletthetStegFelles.evaluerUoppfylt(forsendelseMottatt, brukfrist, AUTO_VENTER_PÅ_KOMPLETT_SØKNAD);
-        }
-        return BehandleStegResultat.utførtUtenAksjonspunkter();
-    }
-
-
-}
