--- behandlingsprosess/src/main/java/no/nav/foreldrepenger/behandling/steg/kompletthet/svp/VurderSøktForTidligStegImpl.java
@@ -1,61 +1,0 @@
-package no.nav.foreldrepenger.behandling.steg.kompletthet.svp;
-
-import static no.nav.foreldrepenger.behandlingslager.behandling.aksjonspunkt.AksjonspunktDefinisjon.VENT_PGA_FOR_TIDLIG_SØKNAD;
-
-import jakarta.enterprise.context.ApplicationScoped;
-import jakarta.inject.Inject;
-
-import no.nav.foreldrepenger.behandling.BehandlingReferanse;
-import no.nav.foreldrepenger.behandling.steg.kompletthet.VurderKompletthetSteg;
-import no.nav.foreldrepenger.behandling.steg.kompletthet.VurderKompletthetStegFelles;
-import no.nav.foreldrepenger.behandlingskontroll.BehandleStegResultat;
-import no.nav.foreldrepenger.behandlingskontroll.BehandlingStegRef;
-import no.nav.foreldrepenger.behandlingskontroll.BehandlingTypeRef;
-import no.nav.foreldrepenger.behandlingskontroll.BehandlingskontrollKontekst;
-import no.nav.foreldrepenger.behandlingskontroll.FagsakYtelseTypeRef;
-import no.nav.foreldrepenger.behandlingslager.behandling.BehandlingStegType;
-import no.nav.foreldrepenger.behandlingslager.behandling.repository.BehandlingRepository;
-import no.nav.foreldrepenger.behandlingslager.fagsak.FagsakYtelseType;
-import no.nav.foreldrepenger.kompletthet.Kompletthetsjekker;
-import no.nav.foreldrepenger.skjæringstidspunkt.SkjæringstidspunktTjeneste;
-
-@BehandlingStegRef(BehandlingStegType.VURDER_KOMPLETT_TIDLIG)
-@BehandlingTypeRef
-@FagsakYtelseTypeRef(FagsakYtelseType.SVANGERSKAPSPENGER)
-@ApplicationScoped
-public class VurderSøktForTidligStegImpl implements VurderKompletthetSteg {
-    private Kompletthetsjekker kompletthetsjekker;
-    private BehandlingRepository behandlingRepository;
-    private SkjæringstidspunktTjeneste skjæringstidspunktTjeneste;
-
-    public VurderSøktForTidligStegImpl() {
-        // CDI
-    }
-
-    @Inject
-    public VurderSøktForTidligStegImpl(Kompletthetsjekker kompletthetsjekker, BehandlingRepository behandlingRepository,
-                                       SkjæringstidspunktTjeneste skjæringstidspunktTjeneste) {
-        this.kompletthetsjekker = kompletthetsjekker;
-        this.behandlingRepository = behandlingRepository;
-        this.skjæringstidspunktTjeneste = skjæringstidspunktTjeneste;
-    }
-
-    @Override
-    public BehandleStegResultat utførSteg(BehandlingskontrollKontekst kontekst) {
-        var behandlingId = kontekst.getBehandlingId();
-        var behandling = behandlingRepository.hentBehandling(behandlingId);
-        var skjæringstidspunkter = skjæringstidspunktTjeneste.getSkjæringstidspunkter(behandlingId);
-        var ref = BehandlingReferanse.fra(behandling);
-
-        if (skalPassereKompletthet(behandling) || kanPassereKompletthet(behandling)) {
-            return BehandleStegResultat.utførtUtenAksjonspunkter();
-        }
-
-        var søknadMottatt = kompletthetsjekker.vurderSøknadMottattForTidlig(ref, skjæringstidspunkter);
-        if (!søknadMottatt.erOppfylt()) {
-            return VurderKompletthetStegFelles.evaluerUoppfylt(søknadMottatt, VENT_PGA_FOR_TIDLIG_SØKNAD);
-        }
-
-        return BehandleStegResultat.utførtUtenAksjonspunkter();
-    }
-}
